move `@ResolvesTypesViaReflection` from package `core` to `base`
This annotation resides in `core` for historic reasons, but it is a very central marker that might as well be placed in `base` to be used by shared utilities.
