+++ archunit/src/main/java/com/tngtech/archunit/core/domain/JavaClassDescriptor.java
@@ -31,7 +31,7 @@ import com.tngtech.archunit.Internal;
 import com.tngtech.archunit.base.ArchUnitException.ReflectionException;
 import com.tngtech.archunit.base.MayResolveTypesViaReflection;
 import com.tngtech.archunit.base.Optional;
-import com.tngtech.archunit.core.ResolvesTypesViaReflection;
+import com.tngtech.archunit.base.ResolvesTypesViaReflection;
 import org.objectweb.asm.Type;
 
 import static com.google.common.base.Preconditions.checkArgument;
