+++ archunit/src/main/java/com/tngtech/archunit/core/domain/JavaConstructor.java
@@ -25,7 +25,7 @@ import com.tngtech.archunit.PublicAPI;
 import com.tngtech.archunit.base.ArchUnitException.InconsistentClassPathException;
 import com.tngtech.archunit.base.MayResolveTypesViaReflection;
 import com.tngtech.archunit.base.Optional;
-import com.tngtech.archunit.core.ResolvesTypesViaReflection;
+import com.tngtech.archunit.base.ResolvesTypesViaReflection;
 import com.tngtech.archunit.core.importer.DomainBuilders.JavaConstructorBuilder;
 
 import static com.google.common.collect.Sets.union;
