+++ archunit/src/main/java/com/tngtech/archunit/core/domain/JavaMethod.java
@@ -26,7 +26,7 @@ import com.tngtech.archunit.base.ArchUnitException.InconsistentClassPathExceptio
 import com.tngtech.archunit.base.Function;
 import com.tngtech.archunit.base.MayResolveTypesViaReflection;
 import com.tngtech.archunit.base.Optional;
-import com.tngtech.archunit.core.ResolvesTypesViaReflection;
+import com.tngtech.archunit.base.ResolvesTypesViaReflection;
 import com.tngtech.archunit.core.importer.DomainBuilders;
 
 import static com.google.common.collect.Sets.union;
